gnhf 10: Extracted a shared src/lib/rng.ts (randInt/shuffle/sample), consolidating three separate Fisher–Yates/sampling implementations across math-gen, quiz-service, and easter-egg into one authoritative module.
